grapherXmpl {deldir}	R Documentation

grapherXmpl

Description

A data set taken from an example in the grapherator package. This data set demonstrates handling a data set with duplicated points.

Usage

grapherXmpl

Format

A data frame with 250 observations on the following 2 variables.

x

a numeric vector

y

a numeric vector

Details

There are 25 duplicated points, so the net number of observations is 225. These data constitute a structure (named coordinates) generated internally in the function addEdgesDelaunay. The call is to be found in the examples in the help file for the plot.grapherator() in the grapherator package. The relevant example initially threw an error, revealing a bug in deldir() that was triggered when there were duplicated points in the data.

Source

The grapherator package, https://CRAN.R-project.org/package=grapherator

Examples

dgX <- deldir(grapherXmpl) # Now works!!!`

[Package deldir version 2.0-4 Index]