print.deldir {deldir}	R Documentation

Print some information about a tessellation/triangulation.

Description

Prints a very brief description of an object of class "deldir" as returned by deldir().

Usage

## S3 method for class 'deldir'
print(x,digits=NULL,...)

Arguments

x
A Delaunay triangulation and Dirichlet (Voronoi) tessellation of a set of points (object of class "deldir").

digits
Integer scalar. The number of digits to which to round the numeric information before printing. Note this may be give negative values. (See round().)

...
Not used.

Details

This is a method for the generic print() function.

Author(s)

Rolf Turner rolfurner@posteo.net

See Also

print()

Examples

set.seed(42)
x    <- rnorm(200,0,4)
y    <- rnorm(200,0,4)
dxy1 <- deldir(x,y)
dxy2 <- deldir(x,y,rw=c(-12,12,-11,11))
dxy1
dxy2
print(dxy1,digits=4)

[Package deldir version 2.0-4 Index]