tileArea {deldir}	R Documentation

Area of a Dirichlet tile.

Description

Calculates the area of a Dirichlet tile, applying a discrete version of Stoke's theorem.

Usage

tileArea(x, y, rw)

Arguments

x
The x-coordinates of the vertices of the tile, in anticlockwise direction. The last coordinate should not repeat the first.

y
The y-coordinates of the vertices of the tile, in anticlockwise direction. The last coordinate should not repeat the first.

rw
A vector of length 4 specifying the rectangular window in which the relevant tessellation was constructed. See deldir() for more detail. Actually this can be any rectangle containing the tile in question.

Details

The heavy lifting is done by the Fortran subroutine stoke() which is called by the .Fortran() function.

Value

A positive scalar.

Author(s)

Rolf Turner rolfurner@posteo.net

See Also

deldir() tilePerim()

Examples

set.seed(42)
x <- runif(20)
y <- runif(20)
z <- deldir(x,y,rw=c(0,1,0,1))
w <- tile.list(z)
with(w[[1]],tileArea(x,y,rw=z$rw))
sapply(w,function(x,rw){tileArea(x$x,x$y,attr(w,"rw"))})
x <- c(0.613102,0.429294,0.386023,0.271880,0.387249,0.455900,0.486101)
y <- c(0.531978,0.609665,0.597780,0.421738,0.270596,0.262953,0.271532)
# The vertices of the Dirichlet tile for point 6.
tileArea(x,y,rw=c(0,1,0,1))
tileArea(x,y,rw=c(-1,2,-3,4)) # Same as above.

[Package deldir version 2.0-4 Index]