which.tile {deldir}	R Documentation

Determine the tile containing a given point.

Description

Finds the Dirichlet/Voronoi tile, of a tessellation produced by deldir(), that contains a given point.

Usage

which.tile(x, y, tl)

Arguments

x
The x coordinate of the point in question.

y
The y coordinate of the point in question.

tl
A tile list, as produced by the function tile.list() from a tessellation produced by deldir().

Details

Just minimises the distance from the point in question to the points of the pattern determining the tiles.

Value

An integer equal to the index of the tile in which the given point lies.

Author(s)

Rolf Turner rolfurner@posteo.net

See Also

tile.list() deldir().

Examples

set.seed(42)
x   <- runif(20,0,100)
y   <- runif(20,0,100)
dxy <- deldir(x,y)
txy <- tile.list(dxy)
i   <- which.tile(30,50,txy) # The value of i here is 14.
plot(txy,showpoints=FALSE)
text(x,y,labels=1:length(txy),col="red")
points(30,50,pch=20,col="blue")

[Package deldir version 2.0-4 Index]