plot.tile.list {deldir}	R Documentation

Plot Dirchlet/Voronoi tiles

Description

A method for plot. Plots (sequentially) the tiles associated with each point in the set being tessellated.

Usage


plot.tile.list(x, verbose = FALSE, close=FALSE, pch=1, polycol=NA,
               showpoints=TRUE, asp=1, ...)
## S3 method for class 'tile.list':
plot(x, verbose = FALSE, close=FALSE, pch=1,
                         polycol=NA, showpoints=TRUE, asp=1, ...)

Arguments

x 	 A list of the tiles in a tessellation, as produced the function tile.list().

verbose 	 Logical scalar; if TRUE the tiles are plotted one at a time (with a ``Go?'' prompt after each) so that the process can be watched.

close 	 Logical scalar; if TRUE the outer edges of of the tiles (i.e. the edges of the enclosing rectangle) are drawn. Otherwise tiles on the periphery of the tessellation are left ``open''.

pch 	 The plotting character for plotting the points of the pattern which was tessellated. Ignored if showpoints is FALSE.

polycol 	 Optional vector of integers (or NAs); the i-th entry indicates with which colour to fill the i-th tile. Note that an NA indicates the use of no colour at all.

showpoints 	 Logical scalar; if TRUE the points of the pattern which was tesselated are plotted.

asp 	 The aspect ratio of the plot; integer scalar or NA. Set this argument equal to NA to allow the data to determine the aspect ratio and hence to make the plot occupy the complete plotting region in both x and y directions. This is inadvisable; see the Warnings.

... 	 Optional arguments; not used. There for consistency with the generic plot function.

Value

NULL; side effect is a plot.

Warnings

The default value for verbose was formerly TRUE; it is now FALSE.

The user is strongly advised not to set the value of asp but rather to leave asp equal to its default value of 1. Any other value distorts the tesselation and destroys the perpendicular appearance of lines which are indeed perpendicular. (And conversely can cause lines which are not perpendicular to appear as if they are.)

The argument asp is present ``just because it can be''.

Author(s)

Rolf Turner r.turner@auckland.ac.nz http://www.math.unb.ca/~rolf

See Also

tile.list()

Examples


  x <- runif(20)
  y <- runif(20)
  z <- deldir(x,y,rw=c(0,1,0,1))
  w <- tile.list(z)
  plot(w)
  ccc <- heat.colors(20) # Or topo.colors(20), or terrain.colors(20)
                         # or cm.colors(20), or rainbox(20).
  plot(w,polycol=ccc,close=TRUE)

[Package deldir version 0.0-7 Index]