tile.centroids {deldir}	R Documentation

Compute centroids of Dirchlet (Voronoi) tiles

Description

Given a list of Dirichlet tiles, as produced by tile.list(), produces a data frame consisting of the centroids of those tiles.

Usage

tile.centroids(xxx)

Arguments

xxx
A list of the tiles (produced by tile.list()) in a Dirichlet tessellation of a set of planar points.

Value

A data frame with two columns named x and y. Each row of this data frame consitutes the centroid of one of the Dirichlet tiles.

Author(s)

Rolf Turner r.turner@auckland.ac.nz

References

URL http://en.wikipedia.org/wiki/Centroid

See Also

tile.list()

Examples

set.seed(42)
x <- runif(20)
y <- runif(20)
d <- deldir(x,y)
l <- tile.list(d)
g <- tile.centroids(l)
## Not run: 
plot(l,close=TRUE)
points(g,pch=20,col="red")

## End(Not run)

[Package deldir version 0.1-12 Index]