tile.list {deldir}	R Documentation

Create a list of tiles in a tessellation

Description

For each point in the set being tessellated produces a list entry describing the Dirichlet/Voronoi tile containing that point.

Usage

 tile.list(object) 

Arguments

object
An object of class deldir as produced by the function deldir().

Value

A list with one entry for each of the points in the set being tesselated. Each entry is in turn a list with components

pt
The coordinates of the point whose tile is being described.

ptType
The “type” of the pt, either “data” or “dummy”. Present only if any dummy points were specified in the call to deldir().

x
The x coordinates of the vertices of the tile, in anticlockwise order.

y
The y coordinates of the vertices of the tile, in anticlockwise order.

bp
Vector of logicals indicating whether the tile vertex is a “real” vertex, or a boundary point, i.e. a point where the tile edge intersects the boundary of the enclosing rectangle.

z
The “auxiliary value” or “weight” associated with the pt; present only if such values were supplied in the call to deldir().

Acknowledgement

The author expresses sincere thanks to Majid Yazdani who found and pointed out a serious bug in tile.list in a previous version (0.0-5) of the deldir package.

Warning

The set of vertices of each tile may be “incomplete”. Only vertices which lie within the enclosing rectangle, and “boundary points” are listed.

Note that the enclosing rectangle may be specified by the user in the call to deldir().

In contrast to some earlier versions of deldir, the corners of the enclosing rectangle are now include as vertices of tiles. I.e. a tile which in fact extends beyond the rectangular window and contains a corner of that window will have that corner added to its list of vertices. Thus the corresponding polygon is the intersection of the tile with the enclosing rectangle.

Author(s)

Rolf Turner r.turner@auckland.ac.nz

See Also

deldir(), plot.tile.list() triang.list() plot.triang.list()

Examples

	x <- runif(20)
	y <- runif(20)
	z <- deldir(x,y)
	w <- tile.list(z)

	z <- deldir(x,y,rw=c(0,1,0,1))
	w <- tile.list(z)

	z <- deldir(x,y,rw=c(0,1,0,1),dpl=list(ndx=2,ndy=2))
	w <- tile.list(z)

[Package deldir version 0.1-12 Index]