tilePerim {deldir}	R Documentation

Calculate tile perimeters.

Description

Calculates the perimeters of all of the Dirichlet (Voronoi) tiles in a tessellation of a set of planar points. Also calculates the sum and the mean of these perimeters.

Usage

tilePerim(object,inclbdry=TRUE)

Arguments

object
An object of class tile.list (as produced by tile.list() specifying the Dirichlet (Voronoi) tiles in a tessellation of a set of planar points.

inclbdry
Logical scalar. Should boundary segments (edges of tiles at least one of whose endpoints lies on the enclosing rectangle rw (see deldir()) be included in the perimeter?

Value

A list with components

perimeters
A vector consisting of the values of the perimeters of the Dirichlet tiles in the tessellation.

totalPerim
The sum of perimeters.

meanPerim
The mean of perimeters.

Note

Function added at the request of Haozhe Zhang.

Author(s)

Rolf Turner r.turner@auckland.ac.nz

See Also

tile.list(), plot.tile.list()

Examples

    x <- runif(20)
    y <- runif(20)
    z <- deldir(x,y,rw=c(0,1,0,1))
    w <- tile.list(z)
    p1 <- tilePerim(w)
    p0 <- tilePerim(w,inclbdry=FALSE)
    p1$totalPerim - p0$totalPerim # Get 4 = the perimeter of rw.
    ss <- apply(as.matrix(z$dirsgs[,1:4]),1,
                function(x){(x[1]-x[3])^2 + (x[2]-x[4])^2})
    2*sum(sqrt(ss)) - p0$totalPerim # Get 0; in tilePerim() each interior
                                    # edge is counted twice.

[Package deldir version 0.1-12 Index]