triang.list {deldir}	R Documentation

Create a list of Delaunay triangles

Description

From an object of class “deldir” produces a list of the Delaunay triangles in the triangulation of a set of points in the plane.

Usage

triang.list(object)

Arguments

object
An object of class “deldir” as produced by deldir().

Value

A list each of whose components is a 3 x 3, 3 x 4 or 3 x 5 data frame corresponding to one of the Delaunay triangles specified by “object”. The rows of each such data frame correspond to the vertices of the corresponding Delaunay triangle. The columns are:

ptNum (the point number of the vertex in the original list of points being triangulated)

ptType (the type of the vertex; “data” or “dummy”; present only if there were any dummy points specified)

x (the x-coordinate of the vertex)

y (the x-coordinate of the vertex)

z (the “auxiliary value” or “weight” z associated with the vertex; present only if such values were supplied in the call to deldir())

The returned value has an attribute “rw” consisting of the enclosing rectangle of the triangulation.

Note

The code of this function was taken more-or-less directly from code written by Adrian Baddeley for the “delaunay()” function in the “spatstat” package.

Author(s)

Rolf Turner r.turner@auckland.ac.nz

See Also

deldir(), plot.triang.list(), tile.list(), plot.tile.list()

Examples

set.seed(42)
x   <- runif(20)
y   <- runif(20)
z   <- sample(1:100,20)
d   <- deldir(x,y,z=z)
ttt <- triang.list(d)

[Package deldir version 0.1-12 Index]